+++ CHANGELOG.md
@@ -11,7 +11,7 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
 
 ### Changed
 
-- MongoDB instrumentation is updated to comply with v1.39.0 Semantic Convention
+- MongoDB instrumentation is updated to comply with v1.39.0 Semantic Convention (for example, `db.system` -> `db.system.name`)
 
 #### Dependency updates